move user & checkup details dto to model.dto
