+++ app/src/main/java/org/maternalcare/modules/main/user/model/dto/UserCheckupDto.kt
@@ -1,4 +1,4 @@
-package org.maternalcare.modules.main.user.model
+package org.maternalcare.modules.main.user.model.dto
 
 data class UserCheckupDto(
     var id: String = "",
